@@ -1,14 +1,14 @@
 name: singleton
 description: Singleton make singleton implementation less hassle, and also make it more friendly to tests.
-version: 0.0.2
+version: 0.0.3-nullsafety
 homepage: https://github.com/timnew/singleton.dart
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  meta: ^1.2.3
-  async: ^2.4.2
+  meta: ^1.3.0-nullsafety.6
+  async: ^2.5.0-nullsafety.3
 
 dev_dependencies:
-  test: ^1.15.4
+  test: ^1.16.0-nullsafety.13
